+++ Editor/Administrator/UITK/CS/PauseWindow.cs
@@ -1,4 +1,4 @@
-﻿using SymphonyFrameWork.System.API;
+﻿using SymphonyFrameWork.System;
 using SymphonyFrameWork.Utility;
 using System.Reflection;
 using System.Threading.Tasks;