+++ Editor/Administrator/UITK/CS/SaveDataRegistryWindow.cs
@@ -1,6 +1,5 @@
 ﻿using SymphonyFrameWork.System.SaveSystem;
 using SymphonyFrameWork.Utility;
-using SaveDataRegistry = SymphonyFrameWork.System.API.SaveDataRegistry;
 using System;
 using System.Collections.Generic;
 using System.Linq;