+++ Editor/PackageInitializer.cs
@@ -2,7 +2,6 @@
 using SymphonyFrameWork.Config;
 using SymphonyFrameWork.Core;
 using SymphonyFrameWork.System.SaveSystem;
-using SaveDataRegistry = SymphonyFrameWork.System.API.SaveDataRegistry;
 using UnityEditor;
 using UnityEngine;