+++ Runtime/Component/SymphonyLocate.cs
@@ -1,5 +1,4 @@
 ﻿using SymphonyFrameWork.System.ServiceLocate;
-using ServiceLocator = SymphonyFrameWork.System.API.ServiceLocator;
 using System;
 using System.Reflection;
 using UnityEngine;