+++ Runtime/Component/SymphonyLocateObject.cs
@@ -1,4 +1,4 @@
-﻿using SymphonyFrameWork.System.API;
+﻿using SymphonyFrameWork.System.ServiceLocate;
 using System.Threading;
 using System.Threading.Tasks;