+++ Samples/Runtime/PauseManagerSample/Scripts/PauseManagerSample_Controller.cs
@@ -1,4 +1,4 @@
-using SymphonyFrameWork.System.API;
+using SymphonyFrameWork.System;
 using System;
 using System.Collections.Generic;
 using System.Text;