+++ Samples/Runtime/SaveDataSystemSample/Scripts/SaveDataSystemSample_Controller.cs
@@ -1,5 +1,4 @@
 ﻿using SymphonyFrameWork.System.SaveSystem;
-using SaveDataRegistry = SymphonyFrameWork.System.API.SaveDataRegistry;
 using System;
 using System.Collections.Generic;
 using System.Linq;