+++ Samples/Runtime/SceneLoaderSample/Scripts/SceneLoaderSample_Controller.cs
@@ -1,5 +1,4 @@
 using SymphonyFrameWork.System.SceneLoad;
-using SceneLoader = SymphonyFrameWork.System.API.SceneLoader;
 using System;
 using System.Collections.Generic;
 using System.Text;