+++ Samples/Runtime/ServiceLocatorSample/Scripts/ServiceLocatorSample_2.cs
@@ -1,4 +1,4 @@
-﻿using SymphonyFrameWork.System.API;
+﻿using SymphonyFrameWork.System.ServiceLocate;
 using System;
 using UnityEngine;